refactor: extract date parsing and formatting functions into date.ts
